+++ CHANGELOG.md
@@ -22,6 +22,9 @@ BUG FIXES:
  * api/client: Fix potentially out of order logs and streamed file contents
    [[GH-4234](https://github.com/hashicorp/nomad/issues/4234)]
 
+BUG FIXES:
+ * driver/docker: Fix docker credential helper support [[GH-4266](https://github.com/hashicorp/nomad/issues/4266)]
+
 ## 0.8.3 (April 27, 2018)
 
 BUG FIXES: